ml-service/similarity/requirements.txt
flask
flask-cors
python-dotenv
pandas
scikit-learn
pymongo
numpy
pyarrow
